BuddysBrain-Backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
sentence-transformers
torch
scikit-learn
scipy
openai>=1.3.5
python-docx
pdfplumber
tiktoken
youtube-transcript-api
cloudinary
python-multipart